requirements.txt
Bottleneck==1.3.2
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
configparser==5.2.0
cryptography==35.0.0
cycler==0.11.0
dataclasses==0.6
docker-pycreds==0.4.0
fonttools==4.25.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
img2vec-pytorch==1.0.1
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
numexpr==2.8.1
numpy==1.21.4
olefile==0.46
packaging==21.3
pandas==1.3.5
pathtools==0.1.2
Pillow==8.4.0
pip==21.2.4
promise==2.3
protobuf==3.15.8
psutil==5.8.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.4
PySocks==1.7.1
python-box==5.4.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
sentry-sdk==1.5.6
setuptools==58.0.4
shortuuid==1.0.8
sip==5.0.0
six==1.16.0
smmap==3.0.5
subprocess32==3.5.4
termcolor==1.1.0
threadpoolctl==2.2.0
torch==1.10.1
torchvision==0.11.2
tornado==6.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.8
wandb==0.12.10
wheel==0.37.1
yaspin==2.1.0
